+++ CHANGELOG.md
@@ -25,11 +25,12 @@
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary.html.twig`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary/summary-pretotal.html.twig`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary/summary-discount.html.twig`
-### 2025-09-07 14:54 — 20e2e2bb
+### 2025-09-07 15:34 — 755384a0
 - Add feature to show discount in cart summary
 
 **Changed files:**
 
-- `custom/plugins/StrixNLUxUpgrades/src/Core/Checkout/Cart/Processor/CartDiscountProcessor.php`
 - `custom/plugins/StrixNLUxUpgrades/src/Resources/config/services.xml`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/page/checkout/summary.html.twig`
+- `custom/plugins/StrixNLUxUpgrades/src/Storefront/Page/Checkout/UnifiedDiscountSummarySubscriber.php`